# bookstore-front/src/app/service/model/book.ts
/**
 * BookStore APIs
 * BookStore APIs exposed from a Java EE back-end to an Angular front-end
 *
 * OpenAPI spec version: 1.0.0
 * Contact: mgeovanny@uninorte.edu.co
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Book resource representation
 */
export class Book {
  /**
   * Identifier
   */
  id?: number;

  /**
   * Title of the book
   */
  title: string;

  /**
   * Summary describing the book
   */
  description?: string;

  /**
   * Unit cost
   */
  unitCost?: number;

  /**
   * ISBN number
   */
  isbn: string;

  /**
   * Date in which the book has been published
   */
  publicationDate?: Date;

  /**
   * Number of pages
   */
  nbOfPages?: number;

  /**
   * URL of the image cover
   */
  imageURL?: string;

  /**
   * Language in which the book has been written
   */
  language?: Book.LanguageEnum;

}
export namespace Book {
  export enum LanguageEnum {
    ENGLISH = <any> 'ENGLISH',
    FRENCH = <any> 'FRENCH',
    SPANISH = <any> 'SPANISH',
    PORTUGUESE = <any> 'PORTUGUESE',
    ITALIAN = <any> 'ITALIAN',
    FINISH = <any> 'FINISH',
    GERMAN = <any> 'GERMAN',
    DUTCH = <any> 'DUTCH',
    RUSSIAN = <any> 'RUSSIAN'
  }
}